requirements.txt
Django==3.2.3
django-admin-interface==0.19.1
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.0
Pillow==9.2.0
PyJWT==2.6.0
python-decouple==3.6